Moved to (global) widget folder for that other pages can use it
